fix: 조회수 타입 변경

+++ app/src/main/java/umc/onairmate/data/model/entity/VideoData.kt
@@ -15,5 +15,5 @@ data class VideoData(
     @SerializedName("videoId")
     val videoId: String,
     @SerializedName("viewCount")
-    val viewCount: Int
+    val viewCount: Long
 )
